backend/requirements.txt
# -------------------------
# Core Framework
# -------------------------
fastapi==0.115.6
uvicorn[standard]==0.32.1

# -------------------------
# Database (Async PostgreSQL)
# -------------------------
SQLAlchemy==2.0.36
asyncpg==0.30.0

# -------------------------
# Auth / Security
# -------------------------
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# -------------------------
# Config / Env
# -------------------------
pydantic-settings==2.6.1
python-dotenv==1.0.1

# -------------------------
# Utilities
# -------------------------
python-multipart==0.0.12
